+++ Example/Tests/TBSMStateMachineSimpleTests.m
@@ -129,7 +129,7 @@ describe(@"TBSMStateMachine", ^{
             TBSMSubState *subStateB = [TBSMSubState subStateWithName:@"subStateB" stateMachine:subStateMachineB];
             subStateMachineA.states = @[subStateB];
             
-            parallelStates.states = @[subStateMachineA];
+            parallelStates.stateMachines = @[subStateMachineA];
             stateMachine.states = @[parallelStates];
             stateMachine.initialState = parallelStates;