+++ Example/Tests/TBSMStateTests.m
@@ -122,7 +122,7 @@ describe(@"TBSMState", ^{
         TBSMSubState *subStateB = [TBSMSubState subStateWithName:@"subStateB" stateMachine:subStateMachineB];
         subStateMachineA.states = @[subStateB];
         
-        parallelStates.states = @[subStateMachineA];
+        parallelStates.stateMachines = @[subStateMachineA];
         stateMachine.states = @[parallelStates];
         stateMachine.initialState = parallelStates;